feat: reversion in db and migration on docker postgres
