+++ package.json
@@ -9,10 +9,6 @@
     "dev": "nodemon src/index.ts",
     "start": "nodemon src/index.ts"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/LiiPeee/tamplate-node-ts.git"
-  },
   "keywords": [],
   "author": "",
   "license": "ISC",
@@ -26,7 +22,6 @@
     "@types/express": "^4.17.21",
     "@types/jest": "^29.5.4",
     "@types/jsonwebtoken": "^9.0.3",
-    "@types/mongoose": "^5.11.97",
     "@types/node": "^20.4.9",
     "jest": "^29.6.4",
     "nodemon": "^3.0.1",
@@ -34,13 +29,13 @@
     "ts-node": "^10.9.2"
   },
   "dependencies": {
-    "@prisma/client": "^5.1.1",
+    "@prisma/client": "^5.9.1",
     "bcrypt": "^5.1.1",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.2",
-    "mongoose": "^8.1.2",
+    "prisma": "^5.9.1",
     "typescript": "^5.3.3"
   }
 }
