+++ prisma/schema.prisma
@@ -1,0 +1,44 @@
+// This is your Prisma schema file,
+// learn more about it in the docs: https://pris.ly/d/prisma-schema
+
+generator client {
+  provider = "prisma-client-js"
+}
+
+datasource db {
+  provider = "postgresql"
+  url      = env("DATABAE_URL")
+}
+
+model Account{
+  id Int @id @default(autoincrement())
+  endDate DateTime?  
+  createDate DateTime @default(now())
+  name String
+  email String @unique
+  balance Float
+  password String
+  token String?  
+  transaction Transaction[]
+}
+
+model Transaction{
+  id Int @id @default(autoincrement())
+  createDate DateTime @default(now())
+  endDate DateTime 
+  value Float
+  formatPayment String
+  paid Boolean @default(false)
+  contacts Contact[]
+  comment String?
+  account Account @relation(fields: [accountId], references: [id])
+  accountId Int
+}
+model Contact{
+  id Int @id @default(autoincrement())
+  creatDate DateTime @default(now())
+  endDate DateTime
+  transaction Transaction[] 
+  name String
+  phone String
+}
