+++ src/domain/entity/account/account.ts
@@ -1,0 +1,13 @@
+import { ITransaction } from "../transaction/transaction";
+
+export interface IAccount {
+  id: number;
+  endDate?: Date;
+  createDate: Date;
+  name: string;
+  email: string;
+  balance?: number;
+  password: string;
+  token?: string | null;
+  transaction?: ITransaction;
+}
