+++ src/domain/entity/contact.ts
@@ -1,0 +1,7 @@
+export interface IContact {
+  id: string;
+  creatDate: Date;
+  endDate: Date;
+  name: string;
+  phone: string;
+}
