--- src/model/category.ts
@@ -1,10 +1,0 @@
-
-export class Category {
-    constructor(
-        public id: number,
-        public createDate: Date,
-        public endDate: Date,
-        public typeTransaction: string,
-        public recurrence: string,
-    ) { }
-}
