--- src/model/contact.ts
@@ -1,9 +1,0 @@
-export class Contact {
-  constructor(
-    public id: number,
-    public creatDate: Date,
-    public endDate: Date,
-    public name: string,
-    public phone: string
-  ) {}
-}
