+++ src/presentation/routes/accounts-routes.ts
@@ -1,10 +1,9 @@
 import * as express from "express";
+import { AccountUseCase } from "../../aplication/usecase/account-use-case";
+import { AccountRepository } from "../../domain/repository/account-repository";
 import { AccountController } from "../controllers/account-controller";
-import { AccountModel } from "../model/account/account";
-import { AccountRepository } from "../repository/account-repository";
-import { AccountUseCase } from "../usecase/account-use-case";
 export const routerAccount = express.Router();
-const accountRespository = new AccountRepository(AccountModel);
+const accountRespository = new AccountRepository();
 const accountUsecase = new AccountUseCase(accountRespository);
 const accountController = new AccountController(accountUsecase);
 
@@ -12,7 +11,7 @@ routerAccount.post(
   "/account",
   accountController.createAccount.bind(accountController)
 );
-routerAccount.get(
-  "/account/:id",
-  accountController.findAccountById.bind(accountController)
-);
+// routerAccount.get(
+//   "/account/:id",
+//   accountController.findAccountById.bind(accountController)
+// );
