+++ tsconfig.json
@@ -1,10 +1,10 @@
 {
   "compilerOptions": {
-    "target": "es2017",
+    "target": "esnext",
     "experimentalDecorators": true,
     "emitDecoratorMetadata": true,
     "module": "commonjs",
-    "rootDir": "./src",
+    "rootDir": "src",
     "outDir": "./dist",
     "removeComments": true,
     "esModuleInterop": true,
@@ -12,5 +12,7 @@
     "strict": true,
     "resolveJsonModule": true,
     "skipLibCheck": true
-  }
+  },
+  "include": ["**/*"],
+  "exclude": ["test/mocks", "node_modules"]
 }
